Email Validator
I am introducing my third project - the Email Validator! This tool allows users to input an email address to validate its authenticity and correctness. It's a handy tool for ensuring data accuracy and reliability. Check it out and validate your emails effortlessly!
In this, I have used an email validator (https://app.emailvalidation.io/dashboard). This is the link to my Dashboard.